Entferne Unterstützung für Wikimedia Commons-Fotos und implementiere AedPhoto-Komponente zur Anzeige von Defibrillator-Fotos mit optimierter Fetch-Logik und Ladeanzeige
