+++ web/src/model/app.ts
@@ -6,21 +6,20 @@ export interface AedData {
   location: string;
   level?: string;
   description?: string;
-  openingHours?: string;
-  operator?: string;
-  operatorPhone?: string;
-  operatorEmail?: string;
+  openingHours?: string; // Opening hours of the AED location
+  operator?: string; // Name of the organization or person responsible for the AED
+  operatorPhone?: string; // Contact phone number for the operator
+  operatorEmail?: string; // Contact e-mail-address for the operator
   access?: 'yes' | 'permissive' | 'private';
   indoor?: 'yes' | 'no';
   latitude: number;
   longitude: number;
-  source?: string;
-  sourceFeature?: Feature;
-  wikimediaCommons?: string;
+  source?: string; // Source of the data, e.g., 'local_knowledge, defikarte.ch'
+  sourceFeature?: Feature; // original feature with eventually additional data
 }
 
 export interface TooltipData {
   title: string;
   content: string;
-  link?: string;
+  link?: string; // Optional link for more information
 }
