+++ web/src/services/wikimedia.service.ts
@@ -1,0 +1,75 @@
+import axios from 'axios';
+
+const COMMONS_API = 'https://commons.wikimedia.org/w/api.php';
+const HTML_TAG_REGEX = /<[^>]+>/g;
+
+export interface WikimediaPhoto {
+  thumbUrl: string;
+  descriptionUrl: string;
+}
+
+function extractFilename(tag: string): string | null {
+  const normalized = tag.trim();
+  if (!normalized) return null;
+
+  if (/^https?:\/\//i.test(normalized)) {
+    try {
+      const url = new URL(normalized);
+      if (url.hostname.toLowerCase() !== 'commons.wikimedia.org') return null;
+      const wikiMatch = url.pathname.match(/^\/wiki\/(File:[^#?]+)/i);
+      if (wikiMatch) return decodeURIComponent(wikiMatch[1]);
+      return null;
+    } catch {
+      return null;
+    }
+  }
+
+  if (/^file:/i.test(normalized)) return normalized;
+  if (/\.(jpe?g|png|gif|webp|svg)$/i.test(normalized)) return `File:${normalized}`;
+  return null;
+}
+
+export const fetchWikimediaPhoto = async (
+  wikimediaCommons?: string,
+  image?: string,
+  signal?: AbortSignal
+): Promise<WikimediaPhoto | null> => {
+  const filename = extractFilename(wikimediaCommons ?? image ?? '');
+  if (!filename) return null;
+
+  try {
+    const response = await axios.get<any>(COMMONS_API, {
+      params: {
+        action: 'query',
+        titles: filename,
+        prop: 'imageinfo',
+        iiprop: 'url|extmetadata',
+        iiurlwidth: '400',
+        format: 'json',
+        origin: '*',
+      },
+      signal,
+    });
+
+    const pages = response.data?.query?.pages ?? {};
+    const page = Object.values(pages)[0] as any;
+    if (!page || page.missing !== undefined) return null;
+
+    const info = page?.imageinfo?.[0];
+    if (!info?.thumburl) return null;
+
+    return {
+      thumbUrl: info.thumburl,
+      descriptionUrl:
+        info.descriptionurl ??
+        `https://commons.wikimedia.org/wiki/${encodeURIComponent(filename)}`,
+    };
+  } catch (error) {
+    if (axios.isCancel(error)) {
+      console.warn('Wikimedia photo request canceled.');
+    } else {
+      console.error(error);
+    }
+    return null;
+  }
+};
